merge tom.brown.code-problem_report_e
refactor problem report. instead of storing the context as a string in problem report keep it as a tuple. when a problem is reported make a new exception that contains a copy of the context. this exception knows how to print itself and the context.

modified feedvalidator to print these exceptions. Now highlights the relevant column in the html. Also added --output option to feedvalidator to control path of output html file. Made -n disable loading the html in a browser, for non-interactive use.