http://codereview.appspot.com/2428
Add EndOfLineChecker, wrapper for a file-like object that checks for consistent
line ends. It checks that each line ends in CRLF or LF, that lines do not
contain other random characters that look like new line separators and that
every line in the file ends with the same end of line sequence. Fixes
http://code.google.com/p/googletransitdatafeed/issues/detail?id=27

Make RecordingProblemReporter stricter. Uses of it now explicitly list every
exception. Fix some bugs where we reported a semantic problem for a field that
was missing or had syntax errors.