googletransit extension of transitfeed.Agency (includes the BCP-47 language code validation).
This change is deprecating field 'agency_ticket_url' in favor of the new field 'agency_fare_url' (https://sites.google.com/site/gtfschanges/spec-changes-summary#agency) and it is deprecating the fields 'agency_lang' and 'agency_timezone' in agency in favor of the fields 'feed_lang' and 'feed_timezone' in feed_info. The agency extension is the first extension which uses the new _DEPRECATED_FIELD_NAMES functionality. So this revision includes some modifications for not having to change all the old unit tests which use one of the deprecated fields. This was done by adding the 'DeprecatedColumn' exception to the ignore_types list of the problem accumulator of most tests. This revision also adds a new command line argument --error_types_ignore_list which allows adding to the ignore_types list e.g. in the integration tests.

This change is adding pybcp47 module for supporting BCP-47 language code validation of the 'agency_lang' and 'feed_lang' fields. The language code parsing is implemented as a separate pybcp47 module for being able to move it later to a separate project if required. It also comes with its own unit tests. These tests can run standalone and they are being invoked from within the testgoogletransitextension.py file for being included in
the nosetests run.

Review at http://codereview.appspot.com/4271073/